+++ app/Http/Controllers/JobController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\StoreJobRequest;
 use App\Http\Requests\UpdateJobRequest;
 use App\Models\Job;
+use App\Models\Invoice;
 
 class JobController extends Controller
 {
@@ -49,6 +50,13 @@ class JobController extends Controller
      */
     public function destroy(Job $job)
     {
+        //Check if there is any invoice with this job:
+        $count = Invoice::where('job_code', $job->code)->count();
+        if ($count > 0) {
+            return response()->json(['message' => 'Job has invoices, cannot be deleted'], 400);
+        }
+
+
         $job->delete();
         return response()->json(['message' => 'Job deleted']);
     }
